refactor: update component structure for improved readability and maintainability
