+++ src/Layout/Header/Header.module.scss
@@ -12,96 +12,5 @@
     position: relative;
     display: flex;
     flex-direction: column;
-
-    .top {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      padding: 24px 0;
-      padding-right: 77px;
-
-      .logo {
-        font-size: 0;
-      }
-
-      ul,
-      nav {
-        display: flex;
-        gap: 77px;
-        align-items: baseline;
-
-        li {
-          height: 27px;
-
-          a {
-            font-family: $font-family;
-            font-size: 22px;
-            font-weight: 400;
-            line-height: 1.4;
-            color: #0c0d0f;
-            text-align: center;
-            letter-spacing: 0.04em;
-          }
-        }
-      }
-    }
-
-    .bottom {
-      position: sticky;
-      z-index: 20;
-      top: 0;
-
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      padding: 24px 45px 24px 19px;
-
-      .wrapper {
-        display: flex;
-        gap: 101px;
-        align-items: center;
-        width: 100%;
-      }
-
-      .icons {
-        display: flex;
-        gap: 32px;
-        align-items: center;
-
-        .cart {
-          &-btn {
-            position: relative;
-
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            padding: 0;
-          }
-
-          &-counter {
-            position: absolute;
-            top: 0;
-            left: 100%;
-            transform: translate(-70%, -50%);
-
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            width: 20px;
-            height: 20px;
-
-            font-size: 0.7em;
-            color: $accent-white;
-
-            background-color: $button-default;
-            border-radius: 50%;
-          }
-        }
-      }
-    }
   }
 }
