+++ src/Layout/Header/Header.tsx
@@ -1,14 +1,14 @@
-import Bottom from './components/Bottom';
-import HeaderTop from './components/Top';
 import styles from './Header.module.scss';
+import HeaderStickyBar from './modules/HeaderStickyBar/HeaderStickyBar';
+import HeaderTopBar from './modules/HeaderTopBar/HeaderTopBar';
 
 const Header = () => {
   return (
     <header className={styles.header}>
       <div className={'container'}>
         <div className={styles['header-inner']}>
-          <HeaderTop />
-          <Bottom />
+          <HeaderTopBar />
+          <HeaderStickyBar />
         </div>
       </div>
     </header>
