+++ src/Layout/Header/modules/HeaderStickyBar/HeaderStickyBar.module.scss
@@ -1,0 +1,58 @@
+@import '@/styles';
+
+.bottom {
+  position: sticky;
+  z-index: 20;
+  top: 0;
+
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  padding: 24px 45px 24px 19px;
+
+  .wrapper {
+    display: flex;
+    gap: 101px;
+    align-items: center;
+    width: 100%;
+  }
+
+  .icons {
+    display: flex;
+    gap: 32px;
+    align-items: center;
+
+    .cart {
+      &-btn {
+        position: relative;
+
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        padding: 0;
+      }
+
+      &-counter {
+        position: absolute;
+        top: 0;
+        left: 100%;
+        transform: translate(-70%, -50%);
+
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        width: 20px;
+        height: 20px;
+
+        font-size: 0.7em;
+        color: $accent-white;
+
+        background-color: $button-default;
+        border-radius: 50%;
+      }
+    }
+  }
+}
