+++ src/Layout/Header/modules/HeaderStickyBar/components/CategoryAll/CategoryAll.module.scss
