+++ src/Layout/Header/modules/HeaderStickyBar/components/CategoryAll/types.ts
